feat(mac): Add System 6.0.8 support with auto-scaling SDK
BREAKING CHANGES:
- PeerTalk_Bootstrap() now auto-scales peer count based on available memory
- Test apps work on both Mac SE (4MB) and Performa 6200 (8MB) with same code

System 6.0.8 Compatibility Fixes:
- PT_Log: Use Create()/GetFInfo()/SetFInfo() instead of PBH* calls
  * PBHCreateSync() fails on System 6.0.8
  * Older File Manager calls work on both MFS and HFS
- Status window: Use vsprintf() instead of vsnprintf() (C99)
- Status window: Use documentProc instead of noGrowDocProc
- Printf format: %zu → %lu with cast for Classic Mac compatibility

SDK Auto-Scaling:
- Detects available memory after MaxApplZone()
- Scales peer count based on memory tiers:
  * < 200KB: 1 peer  (Mac Plus with apps)
  * < 400KB: 2 peers (Mac SE 4MB)
  * < 800KB: 3 peers (Mac SE/30, LC)
  * >= 800KB: Requested count (Mac II, Quadra, PPC)
- Developers just call PeerTalk_Bootstrap(4) - SDK figures out what's possible

Test Results:
Mac SE (4MB RAM, 2 buffers):
  - 256B: 4 KB/s, 512B: 7 KB/s, 1KB: 12 KB/s, 2KB: 17 KB/s (peak)

Performa 6200 (8MB RAM, 4 buffers):
  - 256B: 16 KB/s, 512B: 33 KB/s, 1KB: 64 KB/s, 2KB: 94 KB/s (peak)

Fixes:
- Queue allocation now logs correct size on error
- Text UI working on System 6.0.8 through Mac OS 9

+++ src/core/queue.c
@@ -37,8 +37,8 @@ int pt_queue_init(struct pt_context *ctx, pt_queue *q, uint16_t capacity)
     if (!q->slots) {
         if (ctx) {
             PT_CTX_ERR(ctx, PT_LOG_CAT_MEMORY,
-                      "Failed to allocate %zu bytes for %u queue slots",
-                      alloc_size, capacity);
+                      "Failed to allocate %lu bytes for %u queue slots",
+                      (unsigned long)alloc_size, capacity);
         }
         return PT_ERR_NO_MEMORY;
     }
